+++ Muninn-master-master/app/src/main/java/studio/bachelor/muninn/Muninn.java
@@ -39,11 +39,11 @@ public class Muninn extends Application {
         return context;
     }
 
-    public static int getSizeSetting(int key_id, int default_id) {
+    public static float getSizeSetting(int key_id, int default_id) {
         String default_str = context.getResources().getString(default_id);
         String key_str = context.getResources().getString(key_id);
         String preference = sharedPreferences.getString(key_str, default_str);
-        return Math.abs(Integer.parseInt(preference));
+        return Math.abs(Float.parseFloat(preference));
     }
 
     public static String getColorSetting(int key_id, int default_id) {
